pyproject.toml
[project]
name = "predicting-ball"
version = "1.0.0"
description = "Predicting match and season outcomes in the Bundesliga (and beyond)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.3.1",
    "fastapi>=0.118.3",
    "great-tables>=0.17.0",
    "matplotlib>=3.10.6",
    "numpy>=2.3.1",
    "optuna>=4.5.0",
    "pandas>=2.3.0",
    "pillow>=11.3.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.0",
    "seaborn>=0.13.2",
    "uvicorn[standard]>=0.37.0",
    "streamlit>=1.50.0",
    "altair>=5.5.0",
    "streamlit-aggrid>=1.1.9",
    "watchdog>=6.0.0",
    "streamlit-js-eval>=0.1.7",
    "boto3>=1.40.55",
    "statsmodels>=0.14.5",
    "setuptools>=80.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.5.0",
    "hypothesis>=6.100.0",
    "pandera>=0.20.0",
    "ruff>=0.8.0",
    "ty>=0.0.1a0",
    "pre-commit>=4.0.0",
]

[tool.uv]
environments = [
    "sys_platform == 'linux'",
    "sys_platform == 'darwin'"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]

[tool.hatch.build]
packages = ["src"]

# Testing
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

# Linting
[tool.ruff]
line-length = 95
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "RUF",    # ruff-specific
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument
    "SIM108", # ternary operator
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.isort]
known-first-party = ["src"]

# Type checking
[tool.ty.environment]
python-version = "3.11"
